Moved into adequate folders, made needed changes
